refactor: Correção de eventuais bugs.

+++ src/controllers/products.js
@@ -1,10 +1,12 @@
 const knex = require('../connection');
-const {uploadFile, removeFile} = require('../storage')
+const { uploadFile, removeFile } = require('../storage')
 
 const registerProduct = async (request, response) => {
   try {
     const { descricao, quantidade_estoque, valor, categoria_id } = request.body;
     const produto_imagem = request.file
+
+    const image = produto_imagem.originalname.split(' ').join('-');
 
     const productExist = await knex('produtos')
       .where({ descricao })
@@ -19,23 +21,23 @@ const registerProduct = async (request, response) => {
       .insert({ descricao, quantidade_estoque, valor, categoria_id })
       .returning('*');
 
-      if(produto_imagem){
-        try {
-          const id_product = insertProduct[0].id
+    if (produto_imagem) {
+      try {
+        const id_product = insertProduct[0].id
 
-          const imageProduct = await uploadFile(
-            `products/${id_product}/${produto_imagem.originalname}`,
-            produto_imagem.buffer,
-            produto_imagem.mimetype
-          )
+        const imageProduct = await uploadFile(
+          `products/${id_product}/${image}`,
+          produto_imagem.buffer,
+          produto_imagem.mimetype
+        )
 
-          const updateImage = await knex('produtos').update('produto_imagem', imageProduct.produto_imagem).where('id', id_product).returning('*');
+        const updateImage = await knex('produtos').update('produto_imagem', imageProduct.produto_imagem).where('id', id_product).returning('*');
 
-          return response.status(201).json(updateImage[0]);
-        } catch (error) {
-          return response.status(500).json({ message: 'Erro interno do servidor.' });
-        }
+        return response.status(201).json(updateImage[0]);
+      } catch (error) {
+        return response.status(500).json({ message: 'Erro interno do servidor.' });
       }
+    }
 
     return response.status(201).json(insertProduct[0]);
   } catch (error) {
@@ -49,6 +51,8 @@ const updateProduct = async (request, response) => {
     const { id } = request.params;
     const produto_imagem = request.file;
 
+    const image = produto_imagem.originalname.split(' ').join('-');
+
     const descriptionExist = await knex('produtos')
       .where({ descricao })
       .andWhere('id', '!=', id)
@@ -59,27 +63,30 @@ const updateProduct = async (request, response) => {
       return response.status(409).json({ mensagem: 'Produto já cadastrado.' });
     }
 
-    if(produto_imagem){
+    if (produto_imagem) {
       try {
         const product = await knex('produtos')
           .where({ id })
           .first();
 
 
-        const pathIndex = product.produto_imagem.indexOf(product.id);
-        const path =  product.produto_imagem.slice(pathIndex);
-        await removeFile(`products/${path}`);
+        if (product.produto_imagem !== null) {
+          const pathIndex = product.produto_imagem.indexOf(product.id);
+          const path = product.produto_imagem.slice(pathIndex);
+          await removeFile(`products/${path}`);
+        }
 
         const imageProduct = await uploadFile(
-          `products/${id}/${produto_imagem.originalname}`,
+          `products/${id}/${image}`,
           produto_imagem.buffer,
           produto_imagem.mimetype
         )
 
-        const updateImage = await knex('produtos').update({ descricao, quantidade_estoque, valor, categoria_id, produto_imagem: imageProduct.produto_imagem }).where({id}).returning('*');
+        const updateImage = await knex('produtos').update({ descricao, quantidade_estoque, valor, categoria_id, produto_imagem: imageProduct.produto_imagem }).where({ id }).returning('*');
 
         return response.status(201).json(updateImage[0]);
       } catch (error) {
+        console.log(error.message);
         return response.status(500).json({ message: 'Erro interno do servidor.' });
       }
     }
@@ -99,15 +106,17 @@ const removeProduct = async (request, response) => {
   try {
     const { id } = request.params;
 
-    const sim = await knex('produtos')
+    const productDelete = await knex('produtos')
       .where({ id })
       .delete()
       .returning('*');
 
-    const pathIndex = sim[0].produto_imagem.indexOf(sim[0].id)
-    const path =  sim[0].produto_imagem.slice(pathIndex)
-    await removeFile(`products/${path}`);
-      
+    if (productDelete.produto_imagem !== null) {
+      const pathIndex = productDelete[0].produto_imagem.indexOf(productDelete[0].id)
+      const path = productDelete[0].produto_imagem.slice(pathIndex)
+      await removeFile(`products/${path}`);
+    }
+
     return response
       .status(200)
       .json({ mensagem: 'Produto excluído com sucesso!' });
